fix(fe_section1): update section animations for Sections
