+++ frontend/src/components/sections/Section2.tsx
@@ -1,13 +1,13 @@
-// src/sections/Section1.tsx
+// src/sections/Section2.tsx
 import { motion } from "framer-motion";
-import { useScrollAnime } from "../hooks/useScrollAnime";
+import { useScrollAnime } from "../../hooks/useScrollAnime";
 
 const variants = {
   hidden: { opacity: 0, y: 80 },
   visible: { opacity: 1, y: 0, transition: { duration: 0.8, ease: "easeOut" } },
 };
 
-const Section1 = () => {
+const Section2 = () => {
   const { ref, controls } = useScrollAnime();
 
   return (
@@ -16,11 +16,11 @@ const Section1 = () => {
       initial="hidden"
       animate={controls}
       variants={variants}
-      className="h-screen flex items-center justify-center bg-gradient-to-br from-red-400 to-pink-500"
+      className="h-screen flex items-center justify-center bg-gradient-to-br from-blue-400 to-teal-500"
     >
-      <h1 className="text-white text-6xl font-bold">Section 1</h1>
+      <h1 className="text-white text-6xl font-bold">Section 2</h1>
     </motion.section>
   );
 };
 
-export default Section1;
+export default Section2;
