+++ frontend/src/components/sections/Section3.tsx
@@ -1,10 +1,10 @@
 // src/sections/Section3.tsx
 import { motion } from "framer-motion";
-import { useScrollAnime } from "../hooks/useScrollAnime";
+import { useScrollAnime } from "../../hooks/useScrollAnime";
 
 const variants = {
-  hidden: { opacity: 0, scale: 0.8 },
-  visible: { opacity: 1, scale: 1, transition: { duration: 1, ease: "easeOut" } },
+  hidden: { opacity: 0, y: 80 },
+  visible: { opacity: 1, y: 0, transition: { duration: 0.8, ease: "easeOut" } },
 };
 
 const Section3 = () => {
